# artikelwerk-api/openapi.integration-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Artikelwerk Integrations-API",
    "version": "1.0.0",
    "description": "Verbindlicher Vertrag der aktuell implementierten Generator-API. Nicht implementierte Schreibbereiche stehen ausschließlich unter x-artikelwerk-planned-operations."
  },
  "servers": [{ "url": "/api/integrations/v1" }],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/context": {
      "get": {
        "operationId": "getContext",
        "summary": "Aktive Mandanten, Maßeinheiten und Attribute laden",
        "responses": {
          "200": { "description": "Schreibkontext", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IntegrationContext" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }, "429": { "$ref": "#/components/responses/RateLimited" }, "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/articles/{id}/attributes": {
      "post": {
        "operationId": "setArticleAttribute",
        "summary": "Attributwert setzen oder bei Tag-Attributen ergänzen",
        "parameters": [{ "$ref": "#/components/parameters/ArticleId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SetAttributeInput" } } } },
        "responses": {
          "200": { "description": "Vorhandener Wert geändert", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SetAttributeResponse" } } } },
          "201": { "description": "Wert neu zugewiesen", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SetAttributeResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" }, "429": { "$ref": "#/components/responses/RateLimited" }, "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "Artikelwerk API token (aw_…)" } },
    "parameters": {
      "ArticleId": { "name": "id", "in": "path", "required": true, "description": "Numerische JTL-Artikel-ID als String, nicht die SKU.", "schema": { "type": "string", "minLength": 1, "maxLength": 200 }, "example": "12345" },
...
    },
...
    "responses": {
      "BadRequest": { "description": "Request oder fachlicher Wert ungültig", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Bearer-Token fehlt oder ist ungültig", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "Erforderlicher Scope fehlt", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Artikel oder Vaterartikel nicht gefunden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Conflict": { "description": "SKU, Kontext, Mandant, Idempotency-Key oder parallele Änderung kollidiert", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "RateLimited": { "description": "Mehr als 120 Requests pro Minute für den Token", "headers": { "Retry-After": { "description": "Wartezeit in Sekunden", "schema": { "type": "integer", "minimum": 1 } }, "X-RateLimit-Limit": { "schema": { "type": "integer", "example": 120 } }, "X-RateLimit-Remaining": { "schema": { "type": "integer", "minimum": 0 } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "InternalError": { "description": "Interner oder SQL-Fehler", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "Error": { "type": "object", "required": ["code", "error", "requestId"], "properties": { "code": { "type": "string", "enum": ["BAD_REQUEST", "VALIDATION_ERROR", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "CONFLICT", "PAYLOAD_TOO_LARGE", "RATE_LIMITED", "INTERNAL_ERROR"] }, "error": { "type": "string" }, "requestId": { "type": "string" }, "details": {} }, "example": { "code": "CONFLICT", "error": "Der Idempotency-Key wurde bereits für einen anderen Request verwendet.", "requestId": "req-17" } },
...
      "IntegrationContext": { "type": "object", "required": ["tenants", "units", "attributes"], "properties": { "tenants": { "type": "array", "items": { "$ref": "#/components/schemas/Tenant" } }, "units": { "type": "array", "items": { "$ref": "#/components/schemas/Unit" } }, "attributes": { "type": "array", "items": { "$ref": "#/components/schemas/Attribute" } } }, "example": { "tenants": [{ "id": 4, "name": "CleanYourLeather", "articleCount": 1520, "isDefault": true }], "units": [{ "id": 1, "name": "Milliliter", "code": "ml" }], "attributes": [{ "id": "material", "name": "Material", "dataType": "text", "allowsCustomValue": true }] } },
      "Tenant": { "type": "object", "required": ["id", "name", "articleCount", "isDefault"], "properties": { "id": { "type": "integer", "minimum": 1 }, "name": { "type": "string" }, "articleCount": { "type": "integer", "minimum": 0 }, "isDefault": { "type": "boolean" } } },
      "Unit": { "type": "object", "required": ["id", "name", "code"], "description": "JTL-Maßeinheit. IDs sind installationsabhängig und ausschließlich aus /context zu übernehmen. Artikelwerk führt keine automatische Umrechnung durch.", "properties": { "id": { "type": "integer", "minimum": 0 }, "name": { "type": "string" }, "code": { "type": ["string", "null"] } } },
      "Attribute": { "type": "object", "required": ["id", "name"], "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "dataType": { "type": "string", "enum": ["integer", "decimal", "text", "date"] }, "allowsCustomValue": { "type": "boolean" } } },
...
      "SetAttributeInput": { "type": "object", "additionalProperties": false, "required": ["attributeId", "value"], "properties": { "attributeId": { "type": "string", "minLength": 1, "maxLength": 100 }, "value": { "type": "string", "minLength": 1, "maxLength": 4000 } }, "example": { "attributeId": "material", "value": "Leder" } },
      "SetAttributeResponse": { "type": "object", "required": ["operationId", "articleId", "attributeId", "previousValue", "value", "action"], "properties": { "operationId": { "type": "string", "format": "uuid" }, "articleId": { "type": "string" }, "attributeId": { "type": "string" }, "previousValue": { "type": ["string", "null"] }, "value": { "type": "string" }, "action": { "type": "string", "enum": ["insert", "append", "overwrite"] } }, "example": { "operationId": "f9be70ad-24b9-45b6-9301-32173972f457", "articleId": "12345", "attributeId": "material", "previousValue": null, "value": "Leder", "action": "insert" } },
...
    }
  }
}